[kokkos] Port BeamSpotESProducer + add exception info + move BeamSpotPOD to DataFormats
