+++ src/kokkos/Framework/EventSetup.h
@@ -40,7 +40,7 @@ namespace edm {
           typeToProduct_.emplace(std::type_index(typeid(T)), std::make_unique<ESWrapper<T>>(std::move(prod)));
 #endif
       if (not succeeded.second) {
-        throw std::runtime_error(std::string("Product of type ") + typeid(T).name() + " already exists");
+        throw std::runtime_error(std::string("EventSetup::put(): Product of type ") + typeid(T).name() + " already exists");
       }
     }
 
@@ -48,7 +48,7 @@ namespace edm {
     T const& get() const {
       const auto found = typeToProduct_.find(std::type_index(typeid(T)));
       if (found == typeToProduct_.end()) {
-        throw std::runtime_error(std::string("Product of type ") + typeid(T).name() + " is not produced");
+        throw std::runtime_error(std::string("EventSetup::get(): Product of type ") + typeid(T).name() + " is not produced");
       }
       return static_cast<ESWrapper<T> const&>(*(found->second)).product();
     }
