+++ src/kokkos/Framework/ProductRegistry.h
@@ -29,7 +29,7 @@ namespace edm {
       auto succeeded = typeToIndex_.emplace(ti, Indices{currentModuleIndex_, ind});
 #endif
       if (not succeeded.second) {
-        throw std::runtime_error(std::string("Product of type ") + typeid(T).name() + " already exists");
+        throw std::runtime_error(std::string("ProductRegistry::produces(): Product of type ") + typeid(T).name() + " already exists");
       }
       return EDPutTokenT<T>{ind};
     }
@@ -38,7 +38,7 @@ namespace edm {
     EDGetTokenT<T> consumes() {
       const auto found = typeToIndex_.find(std::type_index(typeid(T)));
       if (found == typeToIndex_.end()) {
-        throw std::runtime_error(std::string("Product of type ") + typeid(T).name() + " is not produced");
+        throw std::runtime_error(std::string("ProductRegistry::consumes(): Product of type ") + typeid(T).name() + " is not produced");
       }
       consumedModules_.insert(found->second.moduleIndex());
       return EDGetTokenT<T>{found->second.productIndex()};
